+++ src/platform/packages/shared/ui-actions-enhanced/dynamic_actions/dynamic_actions_public/README.md
@@ -1,0 +1,3 @@
+# @kbn/ui-actions-enhanced-dynamic-actions-public
+
+Empty package generated by @kbn/generate
