+++ src/platform/packages/shared/ui-actions-enhanced/dynamic_actions/dynamic_actions_public/index.ts
@@ -1,0 +1,21 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+export type { BaseActionFactoryContext } from './src/types';
+export type { ActionFactoryDeps } from './src/action_factory';
+export type { ActionFactoryDefinition } from './src/action_factory_definition';
+export type { State, Transitions, Selectors } from './src/dynamic_action_manager_state';
+export type { ActionStorage } from './src/dynamic_action_storage';
+export type {
+  DynamicActionManagerState,
+  DynamicActionManagerParams,
+} from './src/dynamic_action_manager';
+export { ActionFactory } from './src/action_factory';
+export { DynamicActionManager } from './src/dynamic_action_manager';
+export { AbstractActionStorage, MemoryActionStorage } from './src/dynamic_action_storage';
