+++ src/platform/packages/shared/ui-actions-enhanced/dynamic_actions/dynamic_actions_public/kibana.jsonc
@@ -1,0 +1,7 @@
+{
+  "type": "shared-browser",
+  "id": "@kbn/ui-actions-enhanced-dynamic-actions-public",
+  "owner": "@elastic/appex-sharedux",
+  "group": "platform",
+  "visibility": "shared"
+}
