+++ src/platform/packages/shared/ui-actions-enhanced/dynamic_actions/dynamic_actions_public/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@kbn/ui-actions-enhanced-dynamic-actions-public",
+  "private": true,
+  "version": "1.0.0",
+  "license": "Elastic License 2.0 OR AGPL-3.0-only OR SSPL-1.0"
+}
