+++ src/platform/packages/shared/ui-actions-enhanced/dynamic_actions/dynamic_actions_public/src/action_factory.test.ts
@@ -8,9 +8,9 @@
  */
 
 import { ActionFactory } from './action_factory';
-import { ActionFactoryDefinition } from './action_factory_definition';
+import type { ActionFactoryDefinition } from './action_factory_definition';
 import { licensingMock } from '@kbn/licensing-plugin/public/mocks';
-import { PublicLicense } from '@kbn/licensing-plugin/public';
+import type { PublicLicense } from '@kbn/licensing-plugin/public';
 
 const def: ActionFactoryDefinition = {
   id: 'ACTION_FACTORY_1',
