+++ src/platform/packages/shared/ui-actions-enhanced/dynamic_actions/dynamic_actions_public/src/action_factory.ts
@@ -13,10 +13,8 @@ import type {
   MigrateFunctionsObject,
   GetMigrationFunctionObjectFn,
 } from '@kbn/kibana-utils-plugin/common';
-import type {
-  UiActionsPresentable as Presentable,
-  ActionMenuItemProps,
-} from '@kbn/ui-actions-plugin/public';
+import type { ActionMenuItemProps } from '@kbn/ui-actions-plugin/public';
+import type { Presentable } from '@kbn/ui-actions-browser';
 import type { Configurable } from '@kbn/kibana-utils-plugin/public';
 import type { ILicense, LicensingPluginStart, LicenseType } from '@kbn/licensing-plugin/public';
 import type { UiActionsActionDefinition as ActionDefinition } from '@kbn/ui-actions-plugin/public';
@@ -24,10 +22,10 @@ import type { SavedObjectReference } from '@kbn/core/types';
 import type { PersistableState } from '@kbn/kibana-utils-plugin/common';
 import type {
   BaseActionConfig,
-  BaseActionFactoryContext,
   SerializedAction,
   SerializedEvent,
-} from './types';
+} from '@kbn/ui-actions-enhanced-dynamic-actions-common';
+import type { BaseActionFactoryContext } from './types';
 import type { ActionFactoryDefinition } from './action_factory_definition';
 
 export interface ActionFactoryDeps {
