+++ src/platform/packages/shared/ui-actions-enhanced/dynamic_actions/dynamic_actions_public/src/dynamic_action_enhancement.ts
@@ -8,10 +8,10 @@
  */
 
 import type { SerializableRecord } from '@kbn/utility-types';
-import { EnhancementRegistryDefinition } from '@kbn/embeddable-plugin/public';
-import { SavedObjectReference } from '@kbn/core/types';
-import { DynamicActionsState } from '..';
-import { UiActionsServiceEnhancements } from '../services';
+import type { EnhancementRegistryDefinition } from '@kbn/embeddable-plugin/public';
+import type { SavedObjectReference } from '@kbn/core/types';
+import type { DynamicActionsState } from '@kbn/ui-actions-enhanced-dynamic-actions-common';
+import type { UiActionsServiceEnhancements } from '@kbn/ui-actions-enhanced-plugin/public';
 
 export const dynamicActionEnhancement = (
   uiActionsEnhanced: UiActionsServiceEnhancements
