+++ src/platform/packages/shared/ui-actions-enhanced/dynamic_actions/dynamic_actions_public/src/dynamic_action_grouping.ts
@@ -9,7 +9,7 @@
 
 import { i18n } from '@kbn/i18n';
 import type { EmbeddableApiContext } from '@kbn/presentation-publishing';
-import { UiActionsPresentableGrouping as PresentableGrouping } from '@kbn/ui-actions-plugin/public';
+import type { PresentableGrouping } from '@kbn/ui-actions-browser';
 
 export const dynamicActionGrouping: PresentableGrouping<EmbeddableApiContext> = [
   {
