+++ src/platform/packages/shared/ui-actions-enhanced/dynamic_actions/dynamic_actions_public/src/dynamic_action_manager.ts
@@ -7,16 +7,20 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import { createStateContainer, StateContainer } from '@kbn/kibana-utils-plugin/common';
-import { EmbeddableApiContext } from '@kbn/presentation-publishing';
-import { UiActionsActionDefinition as ActionDefinition } from '@kbn/ui-actions-plugin/public';
-import { Subscription } from 'rxjs';
+import type { StateContainer } from '@kbn/kibana-utils-plugin/common';
+import type { EmbeddableApiContext } from '@kbn/presentation-publishing';
+import type { UiActionsActionDefinition as ActionDefinition } from '@kbn/ui-actions-plugin/public';
+import type { Subscription } from 'rxjs';
+import type {
+  SerializedAction,
+  SerializedEvent,
+} from '@kbn/ui-actions-enhanced-dynamic-actions-common';
+import { createStateContainer } from '@kbn/kibana-utils-plugin/common';
 import { v4 as uuidv4 } from 'uuid';
-import { StartContract } from '../plugin';
+import type { AdvancedUiActionsStart } from '@kbn/ui-actions-enhanced-plugin/public';
+import type { ActionStorage } from './dynamic_action_storage';
 import { dynamicActionGrouping } from './dynamic_action_grouping';
 import { defaultState, selectors, State, transitions } from './dynamic_action_manager_state';
-import { ActionStorage } from './dynamic_action_storage';
-import { SerializedAction, SerializedEvent } from './types';
 
 const compareEvents = (
   a: ReadonlyArray<{ eventId: string }>,
@@ -32,7 +36,7 @@ export type DynamicActionManagerState = State;
 export interface DynamicActionManagerParams {
   storage: ActionStorage;
   uiActions: Pick<
-    StartContract,
+    AdvancedUiActionsStart,
     | 'registerAction'
     | 'attachAction'
     | 'unregisterAction'
