+++ src/platform/packages/shared/ui-actions-enhanced/dynamic_actions/dynamic_actions_public/src/types.ts
@@ -7,9 +7,9 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export type * from './types';
-export * from './action_factory';
-export type * from './action_factory_definition';
-export * from './dynamic_action_storage';
-export * from './dynamic_action_manager_state';
-export * from './dynamic_action_manager';
+/**
+ * Action factory context passed into ActionFactories' CollectConfig, getDisplayName, getIconType
+ */
+export interface BaseActionFactoryContext {
+  triggers: string[];
+}
