+++ src/platform/plugins/shared/ui_actions_enhanced/public/components/action_factory_picker/action_factory_picker.tsx
@@ -8,7 +8,10 @@
  */
 
 import React from 'react';
-import { ActionFactory, BaseActionFactoryContext } from '../../dynamic_actions';
+import {
+  ActionFactory,
+  BaseActionFactoryContext,
+} from '@kbn/ui-actions-enhanced-dynamic-actions-public';
 import { PresentablePicker, Item } from '../presentable_picker';
 
 export interface ActionFactoryPickerProps {
