+++ src/platform/plugins/shared/ui_actions_enhanced/public/components/action_wizard/test_data.tsx
@@ -17,7 +17,7 @@ import {
 } from '@kbn/ui-actions-plugin/public';
 import { APPLY_FILTER_TRIGGER } from '@kbn/data-plugin/public';
 import { SELECT_RANGE_TRIGGER, VALUE_CLICK_TRIGGER } from '@kbn/embeddable-plugin/public';
-import { ActionFactory, ActionFactoryDefinition, BaseActionConfig } from '../../dynamic_actions';
+import { ActionFactory, ActionFactoryDefinition, BaseActionConfig } from '../../../../../../packages/shared/ui-actions-enhanced/dynamic_actions/dynamic_actions_public/src';
 import { ActionWizard } from './action_wizard';
 
 export const dashboards = [
