+++ src/platform/plugins/shared/ui_actions_enhanced/public/index.ts
@@ -20,23 +20,7 @@ export type {
   StartContract as AdvancedUiActionsStart,
 } from './plugin';
 
-export type {
-  ActionFactoryDefinition as UiActionsEnhancedActionFactoryDefinition,
-  SerializedAction as UiActionsEnhancedSerializedAction,
-  SerializedEvent as UiActionsEnhancedSerializedEvent,
-  DynamicActionManagerParams as UiActionsEnhancedDynamicActionManagerParams,
-  DynamicActionManagerState as UiActionsEnhancedDynamicActionManagerState,
-  BaseActionFactoryContext as UiActionsEnhancedBaseActionFactoryContext,
-  BaseActionConfig as UiActionsEnhancedBaseActionConfig,
-} from './dynamic_actions';
-export {
-  ActionFactory as UiActionsEnhancedActionFactory,
-  AbstractActionStorage as UiActionsEnhancedAbstractActionStorage,
-  DynamicActionManager as UiActionsEnhancedDynamicActionManager,
-  MemoryActionStorage as UiActionsEnhancedMemoryActionStorage,
-} from './dynamic_actions';
-
-export type { DynamicActionsState } from './services/ui_actions_service_enhancements';
+export { UiActionsServiceEnhancements } from './services';
 
 export type {
   DrilldownDefinition as UiActionsEnhancedDrilldownDefinition,
