+++ src/platform/plugins/shared/ui_actions_enhanced/public/services/ui_actions_service_enhancements.ts
@@ -13,15 +13,17 @@ import { ILicense } from '@kbn/licensing-plugin/common/types';
 import { LicensingPluginSetup, LicensingPluginStart } from '@kbn/licensing-plugin/public';
 import { SavedObjectReference } from '@kbn/core/types';
 import { PersistableStateDefinition } from '@kbn/kibana-utils-plugin/common';
-import { DynamicActionsState } from '@kbn/ui-actions-enhanced-dynamic-actions-common';
-import { DrilldownDefinition } from '../drilldowns';
+import {
+  DynamicActionsState,
+  BaseActionConfig,
+  SerializedEvent,
+} from '@kbn/ui-actions-enhanced-dynamic-actions-common';
 import {
   ActionFactory,
   ActionFactoryDefinition,
-  BaseActionConfig,
   BaseActionFactoryContext,
-  SerializedEvent,
-} from '../dynamic_actions';
+} from '@kbn/ui-actions-enhanced-dynamic-actions-public';
+import { DrilldownDefinition } from '../drilldowns';
 import { ActionFactoryRegistry } from '../types';
 
 export type { DynamicActionsState };
