+++ x-pack/platform/plugins/private/drilldowns/url_drilldown/tsconfig.json
@@ -20,7 +20,9 @@
     "@kbn/core-ui-settings-browser-mocks",
     "@kbn/core-ui-settings-browser",
     "@kbn/core-theme-browser-mocks",
-    "@kbn/presentation-publishing"
+    "@kbn/presentation-publishing",
+    "@kbn/ui-actions-enhanced-dynamic-actions-common",
+    "@kbn/ui-actions-enhanced-dynamic-actions-public"
   ],
   "exclude": [
     "target/**/*",
