+++ x-pack/platform/plugins/shared/dashboard_enhanced/public/services/drilldowns/actions/flyout_create_drilldown/flyout_create_drilldown.test.tsx
@@ -8,13 +8,13 @@
 import { BehaviorSubject } from 'rxjs';
 import { ActionDefinitionContext } from '@kbn/ui-actions-plugin/public/actions';
 import {
-  UiActionsEnhancedMemoryActionStorage as MemoryActionStorage,
-  UiActionsEnhancedDynamicActionManager as DynamicActionManager,
-} from '@kbn/ui-actions-enhanced-plugin/public';
+  MemoryActionStorage,
+  DynamicActionManager,
+} from '@kbn/ui-actions-enhanced-dynamic-actions-public';
 import type { EmbeddableApiContext, ViewMode } from '@kbn/presentation-publishing';
 import { flyoutCreateDrilldownAction } from './flyout_create_drilldown';
 import { uiActionsEnhancedPluginMock } from '@kbn/ui-actions-enhanced-plugin/public/mocks';
-import { UiActionsEnhancedActionFactory } from '@kbn/ui-actions-enhanced-plugin/public';
+import { ActionFactory as UiActionsEnhancedActionFactory } from '@kbn/ui-actions-enhanced-dynamic-actions-public';
 import { DynamicActionsSerializedState } from '@kbn/embeddable-enhanced-plugin/public';
 
 jest.mock('../../../kibana_services', () => {
