+++ x-pack/platform/plugins/shared/embeddable_enhanced/tsconfig.json
@@ -13,6 +13,8 @@
     "@kbn/kibana-utils-plugin",
     "@kbn/data-plugin",
     "@kbn/presentation-publishing",
+    "@kbn/ui-actions-enhanced-dynamic-actions-common",
+    "@kbn/ui-actions-enhanced-dynamic-actions-public",
   ],
   "exclude": [
     "target/**/*",
