+++ x-pack/platform/plugins/shared/lens/tsconfig.json
@@ -125,6 +125,7 @@
     "@kbn/deeplinks-analytics",
     "@kbn/core-http-server",
     "@kbn/presentation-util",
+    "@kbn/ui-actions-enhanced-dynamic-actions-public",
   ],
   "exclude": ["target/**/*"]
 }
